move CustomTagProvider to distributeme-core
