+++ distributeme-consul-registry-connector/src/main/java/org/distributeme/consulintegration/ConsulRegistryConnector.java
@@ -9,10 +9,7 @@ import com.google.gson.Gson;
 import com.sun.jersey.api.client.Client;
 import com.sun.jersey.api.client.ClientResponse;
 import com.sun.jersey.api.client.WebResource;
-import org.distributeme.core.Location;
-import org.distributeme.core.RegistryConnector;
-import org.distributeme.core.RegistryLocation;
-import org.distributeme.core.ServiceDescriptor;
+import org.distributeme.core.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
