+++ distributeme-consul-registry-connector/src/test/java/org/distributeme/consulintegration/DistributeMeCustomTagTestClassA.java
@@ -1,5 +1,7 @@
 package org.distributeme.consulintegration;
 
+import org.distributeme.core.CustomTagProvider;
+
 /**
  * @author bjochheim
  * @since 3/17/17.
