+++ distributeme-core/src/main/java/org/distributeme/core/CustomTagProvider.java
@@ -1,4 +1,4 @@
-package org.distributeme.consulintegration;
+package org.distributeme.core;
 
 /**
  * @author bjochheim
